# Perf: Array#at is slow for a hot function like `AstPath.node`