@fisker 2023-02-22
Thanks for this, as I tested, indeed `.at` is  slow.

```console
> console.time('length-1');for(let i = 0; i < 1e8;i++)array[array.length-1];console.timeEnd('length-1')
length-1: 90.51ms
undefined
> console.time('at');for(let i = 0; i < 1e8;i++)array.at(-1);console.timeEnd('at')
at: 5.378s
undefined
```