@fisker 2023-02-22
It's strange that it's not slow in chrome.