@fisker 2023-02-22
It's also fast on Node.js v19.