@Naddiseo 2023-02-22
> It's also fast in Node.js v19.

I guess I'm using node v16, and I'll be stuck using it for a while. It's possible that newer node versions have since optimized the `.at(-1)` codepath.